+++ packages/ui-components/src/components/DateDisplay/index.js
@@ -1,0 +1,2 @@
+export * from './DiagnosticInfo';
+export * from './DateDisplay';
