+++ packages/ui-components/src/components/index.js
@@ -12,3 +12,4 @@ export * from './CustomToastContainer';
 export * from './Suggester';
 export * from './DateDisplay';
 export * from './PatientNameDisplay';
+export * from './DateDisplay'
